+++ functions.php
@@ -61,6 +61,7 @@ require_once(\get_template_directory() . '/plugins/helper/EdkKillboardHelper.php
 require_once(\get_template_directory() . '/plugins/helper/ZkbKillboardHelper.php');
 require_once(\get_template_directory() . '/plugins/widgets/KillboardWidget.php');
 require_once(\get_template_directory() . '/plugins/Killboard.php');
+require_once(\get_template_directory() . '/plugins/ChildpageMenu.php');
 require_once(\get_template_directory() . '/plugins/LatestBlogPosts.php');
 require_once(\get_template_directory() . '/plugins/EveOnlineAvatar.php');
 require_once(\get_template_directory() . '/plugins/HtmlMinify.php');
@@ -98,6 +99,7 @@ new Plugins\Corppage;
 new Plugins\EncodeEmailAddresses;
 new Plugins\Whitelabel;
 new Plugins\Killboard(true);
+new Plugins\ChildpageMenu;
 new Plugins\LatestBlogPosts;
 new Plugins\EveOnlineAvatar;